requirements.txt
beautifulsoup4==4.9.1
dill==0.3.2
fastapi==0.58.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
python-multipart==0.0.5
requests==2.24.0
starlette==0.13.4
tables==3.6.1
tensorflow==2.2.0
tensorflow-hub==0.8.0
tensorflow-text==2.2.1
twilio==6.43.0
uvicorn==0.11.5
wandb==0.9.2
